Major changes to CSS, HTML, and Javascript files
Added a new  new order button., that refreshes the window. Added a  icon. formatted the code to make it easier to read.  Moved where the buttons appear on the dom.

For CSS completely changed the look and feel of the website. Added gird layout. Adding work in progress media queries for smartphone and tablet sizes.

For JavaScript I removed the console.logs. Added some of the information that was outputted to the console and placed it so the user can see on the webpage. Added a function that refreshes the page when the new order button is pressed. Changed the concatenation to a template literal.
Formatted the document to make it more readable.